feat (chonk): upgrade astro
This was a nightmare and I'm still not totally sure what the point was. We have all of the JSON files in src/data instead of src/content because the astro/loaders:glob import does not work inside src/content. What?

We needed to add a generateId function when defining the loaders. I dunno what the default is - afaik this is not documented, but I don't care to dig further.